+++ dao/index.js
@@ -26,4 +26,5 @@ module.exports = {
   tag_dao: loadDao("tag"), // 标签
   goods_dao: loadDao("goods"), // 产出
   produce_dao: loadDao("produce"), // 产品
+  notice_dao: loadDao("notice"), // 公告
 };