+++ dao/notice.js
@@ -1,0 +1,24 @@
+/**
+  * afterloe - cynomy_portal_server/dao/notice.js
+  *
+  * Copyright(c) afterloe.
+  * MIT Licensed
+  *
+  * Authors:
+  *   afterloe <lm6289511@gmail.com> (https://github.com/afterloe)
+  * Date:
+  *   2017-3-9 15:32:43
+  */
+"use strict";
+
+const {resolve} = require("path");
+const commonsLib = require(resolve(__dirname, "public"))();
+
+const classMethod = {
+};
+
+Object.assign(commonsLib, classMethod);
+
+const className = "notice";
+
+module.exports = _ => _.definition({classMethod: commonsLib, className});